use GPU for LPPLSolver as well

+++ src/main/java/kic/lppl/LPPLKernel.java
@@ -3,24 +3,24 @@ package kic.lppl;
 import com.aparapi.Kernel;
 
 public class LPPLKernel extends Kernel {
-    private double[] values;     // FIXME change LevenbergMarquardt to work with floats by overriding the derivatives function
-    private double[] abcc, mwtc; // FIXME change LevenbergMarquardt to work with floats by overriding the derivatives function
+    private float[] values;
+    private float[] abcc, mwtc;
     private float[] time;
 
     public LPPLKernel(float[] time) {
         setExplicit(true);
         this.time = time;
-        this.values = new double[time.length];
+        this.values = new float[time.length];
         put(this.time).put(this.values);
     }
 
     public void setNewTime(float[] time) {
         this.time = time;
-        this.values = new double[time.length];
+        this.values = new float[time.length];
         put(this.time).put(this.values);
     }
 
-    public void setAbccMwtc(double[] abcc, double[] mwtc) {
+    public void setAbccMwtc(float[] abcc, float[] mwtc) {
         this.abcc = abcc;
         this.mwtc = mwtc;
         put(this.abcc).put(mwtc);
@@ -29,16 +29,16 @@ public class LPPLKernel extends Kernel {
     @Override
     public void run() {
         int i = getGlobalId();
-        double t = time[i];
-        double w = mwtc[1];
-        double tc_t = mwtc[2] - t;
-        double a = pow((tc_t), mwtc[0]);
+        float t = time[i];
+        float w = mwtc[1];
+        float tc_t = mwtc[2] - t;
+        float a = pow((tc_t), mwtc[0]);
 
         // return A + B * a + C1 * a * cos(w * log(tc - t)) + C2 * a * sin(w * log(tc - t))
         values[i] =  abcc[0] + abcc[1] * a + abcc[2] * a * cos(w * log(tc_t)) + abcc[3] * a * sin(w * log(tc_t));
     }
 
-    public double[] getValues() {
+    public float[] getValues() {
         get(this.values);
         return values;
     }
